database/requirements.txt
firebase-admin
geopy
openpyxl
pandas
python-dotenv